+++ .hemtt/launch.toml
@@ -2,7 +2,7 @@
 workshop = [
     "450814997",  # CBA_A3's Workshop ID
     "463939057",  # ACE3
-    "2369477168", # Advanced Developer Tools
+#    "2369477168", # Advanced Developer Tools
     "1779063631", # ZEN
     "3355988282", # CVO Media
 ]
@@ -14,6 +14,8 @@ parameters = [
     "-window",
     "-noPause",
     "-noPauseAudio",
+    "-mod=zrn\\adt",
+
 ]
 
 [debugger]
