+++ .hemtt/missions/exec.VR/mission.sqm
@@ -8,14 +8,14 @@ class EditorData
 	toggles=517;
 	class ItemIDProvider
 	{
-		nextID=189;
+		nextID=204;
 	};
 	class Camera
 	{
-		pos[]={-0.67291784,48.773838,-76.524162};
-		dir[]={0.074861161,-0.53466284,0.84174311};
-		up[]={0.047363766,0.84506547,0.53256088};
-		aside[]={0.99606854,3.7252903e-09,-0.088586226};
+		pos[]={66.527977,11.669208,-39.603714};
+		dir[]={-0.62873197,-0.35711682,0.69077033};
+		up[]={-0.2403816,0.93405974,0.26410058};
+		aside[]={0.73953551,0,0.67311752};
 	};
 };
 binarizationWanted=0;
@@ -25,13 +25,14 @@ addons[]=
 	"A3_Characters_F",
 	"A3_Modules_F_Curator_Curator",
 	"A3_Soft_F_Quadbike_01",
-	"A3_Structures_F_Kart_Mil_Flags"
+	"A3_Structures_F_Kart_Mil_Flags",
+	"cvo_branding"
 };
 class AddonsMetaData
 {
 	class List
 	{
-		items=4;
+		items=5;
 		class Item0
 		{
 			className="A3_Characters_F";
@@ -60,11 +61,19 @@ class AddonsMetaData
 			author="Bohemia Interactive";
 			url="https://www.arma3.com";
 		};
+		class Item4
+		{
+			className="cvo_branding";
+			name="CVO Auxiliary - Branding";
+			author="CVO Team";
+			url="https://github.com/OverlordZorn/CVO-Auxiliary";
+		};
 	};
 };
 dlcs[]=
 {
-	"Kart"
+	"Kart",
+	"Aow"
 };
 randomSeed=4540948;
 class ScenarioData
@@ -94,7 +103,7 @@ class Mission
 	};
 	class Entities
 	{
-		items=44;
+		items=52;
 		class Item0
 		{
 			dataType="Group";
@@ -2708,6 +2717,126 @@ class Mission
 			};
 			id=188;
 			type="Flag_Fuel_F";
+		};
+		class Item44
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={58,8.9768047,-22};
+			};
+			side="Empty";
+			flags=5;
+			class Attributes
+			{
+			};
+			id=192;
+			type="cvo_Flag_Olive";
+		};
+		class Item45
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={44,5.6528969,-22};
+			};
+			side="Empty";
+			flags=4;
+			class Attributes
+			{
+			};
+			id=193;
+			type="cvo_Leadership_Portrait";
+		};
+		class Item46
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={48,6,-22};
+			};
+			side="Empty";
+			class Attributes
+			{
+			};
+			id=195;
+			type="cvo_stencil_black";
+			atlOffset=1;
+		};
+		class Item47
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={50,6,-22};
+			};
+			side="Empty";
+			class Attributes
+			{
+			};
+			id=196;
+			type="cvo_stencil_red";
+			atlOffset=1;
+		};
+		class Item48
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={46,6,-22};
+			};
+			side="Empty";
+			class Attributes
+			{
+			};
+			id=197;
+			type="cvo_stencil_white";
+			atlOffset=1;
+		};
+		class Item49
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={56,8.9768047,-22};
+			};
+			side="Empty";
+			flags=5;
+			class Attributes
+			{
+			};
+			id=198;
+			type="cvo_Flag_Olive";
+		};
+		class Item50
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={56,5.9885817,-26};
+			};
+			side="Empty";
+			flags=4;
+			class Attributes
+			{
+			};
+			id=202;
+			type="cvo_carrier_flag_olive";
+		};
+		class Item51
+		{
+			dataType="Object";
+			class PositionInfo
+			{
+				position[]={58,5.9885817,-26};
+			};
+			side="Empty";
+			flags=4;
+			class Attributes
+			{
+			};
+			id=203;
+			type="cvo_carrier_flag_red";
 		};
 	};
 };
