+++ addons/branding/CfgWeapons.hpp
@@ -1,0 +1,3 @@
+class CfgWeapons {
+    #include "CfgWeapons_ace_flags.hpp"
+};
