+++ addons/branding/config.cpp
@@ -44,5 +44,6 @@ class CfgPatches {
 
 #include "CfgVehicles.hpp"
 #include "CfgMarkers.hpp"
+#include "CfgWeapons.hpp"
 
 #include "ACE_Tags.hpp"
