+++ addons/main/script_version.hpp
@@ -1,7 +1,7 @@
 #define MAJOR 1
 #define MINOR 4
 #define PATCH 0
-#define BUILD 300
+#define BUILD 309
 
 
 // #define VERSION MACROS
